# book-network-ui/src/app/services/models/book-create-request.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface BookCreateRequest {
  authorName: string;
  isbn: string;
  shareable?: boolean;
  synopsis: string;
  title: string;
}